refactor: move `Day` struct to `template` module
