+++ src/lib.rs
@@ -1,4 +1,1 @@
-mod day;
 pub mod template;
-
-pub use day::*;
