+++ src/template/commands/all.rs
@@ -1,10 +1,10 @@
 use std::io;
 
 use crate::template::{
+    all_days,
     readme_benchmarks::{self, Timings},
-    ANSI_BOLD, ANSI_ITALIC, ANSI_RESET,
+    Day, ANSI_BOLD, ANSI_ITALIC, ANSI_RESET,
 };
-use crate::{all_days, Day};
 
 pub fn handle(is_release: bool, is_timed: bool) {
     let mut timings: Vec<Timings> = vec![];
@@ -65,7 +65,7 @@ pub fn get_path_for_bin(day: Day) -> String {
 /// This module encapsulates interaction with these binaries, both invoking them as well as parsing the timing output.
 mod child_commands {
     use super::{get_path_for_bin, Error};
-    use crate::Day;
+    use crate::template::Day;
     use std::{
         io::{BufRead, BufReader},
         path::Path,
